@@ -8,7 +8,7 @@ import (
 	"github.com/SSHZ-ORG/tree-diagram/models"
 	"github.com/SSHZ-ORG/tree-diagram/pb"
 	"github.com/SSHZ-ORG/tree-diagram/utils"
-	"google.golang.org/appengine/log"
+	"google.golang.org/appengine/v2/log"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
