@@ -1,6 +1,7 @@
 package models
 
 import (
+	"context"
 	"time"
 
 	"cloud.google.com/go/civil"
@@ -8,9 +9,8 @@ import (
 	"github.com/SSHZ-ORG/tree-diagram/utils"
 	"github.com/pkg/errors"
 	"github.com/qedus/nds"
-	"golang.org/x/net/context"
-	"google.golang.org/appengine/datastore"
-	"google.golang.org/appengine/log"
+	"google.golang.org/appengine/v2/datastore"
+	"google.golang.org/appengine/v2/log"
 	"google.golang.org/protobuf/proto"
 )
 
