@@ -2,7 +2,7 @@ package models
 
 import (
 	"github.com/scylladb/go-set/strset"
-	"google.golang.org/appengine/datastore"
+	"google.golang.org/appengine/v2/datastore"
 )
 
 const (
