@@ -1,13 +1,13 @@
 package scheduler
 
 import (
+	"context"
 	"net/url"
 	"strconv"
 
 	"github.com/SSHZ-ORG/tree-diagram/paths"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
-	"google.golang.org/appengine/taskqueue"
+	"google.golang.org/appengine/v2/taskqueue"
 )
 
 const actorQueueName = "actor-queue"
